screen transition and disabling the welcome screen for the time being
